userstate: move sync API handler into userstate package
Inline signed-field slice, drop dedicated SyncRateLimiter var, and move the handler to userstate/api.go behind a thin auth wrapper.
