+++ userstate/api.go
@@ -1,4 +1,4 @@
-package main
+package userstate
 
 import (
 	"encoding/json"
@@ -6,11 +6,14 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/mtgban/mtgban-website/userstate"
+	"github.com/mtgban/mtgban-website/ratelimit"
 )
 
-// maxUserStateBody is the server-side cap on a sync payload.
-const maxUserStateBody = 512 * 1024
+// maxBody is the server-side cap on a sync payload.
+const maxBody = 512 * 1024
+
+// apiLimiter throttles /api/userstate/ separately from page rendering.
+var apiLimiter = ratelimit.NewLimiter(10, 5) // 10 req/s, burst 5
 
 // writeJSON is a small helper for JSON responses.
 func writeJSON(w http.ResponseWriter, status int, v any) {
@@ -19,23 +22,19 @@ func writeJSON(w http.ResponseWriter, status int, v any) {
 	json.NewEncoder(w).Encode(v)
 }
 
-// UserStateAPI serves GET/PUT on /api/userstate/ and PATCH on /{section}; anonymous is 401, unconfigured DB is 503.
-func UserStateAPI(w http.ResponseWriter, r *http.Request) {
-	email := signedUserEmail(r)
-	if email == "" {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "not signed in"})
-		return
-	}
-	if !SyncRateLimiter.Allow(email) {
+// ServeAPI handles GET/PUT on /api/userstate/ and PATCH on /{section} for an
+// already-authenticated email; a nil db is 503.
+func ServeAPI(w http.ResponseWriter, r *http.Request, db *Client, email string) {
+	if !apiLimiter.Allow(email) {
 		writeJSON(w, http.StatusTooManyRequests, map[string]string{"error": "too many requests"})
 		return
 	}
-	if UserStateDB == nil {
+	if db == nil {
 		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "sync unavailable"})
 		return
 	}
 
-	hash := userstate.HashEmail(email)
+	hash := HashEmail(email)
 	ctx := r.Context()
 
 	// Determine section suffix after the base path, if any.
@@ -50,13 +49,13 @@ func UserStateAPI(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Cache-Control", "private, no-store")
 		// Conditional GET: matching version returns 304 without reading the row.
 		if inm := strings.Trim(r.Header.Get("If-None-Match"), `"`); inm != "" {
-			if v, verr := UserStateDB.GetVersion(ctx, hash); verr == nil && strconv.FormatInt(v, 10) == inm {
+			if v, verr := db.GetVersion(ctx, hash); verr == nil && strconv.FormatInt(v, 10) == inm {
 				w.Header().Set("ETag", `"`+inm+`"`)
 				w.WriteHeader(http.StatusNotModified)
 				return
 			}
 		}
-		st, err := UserStateDB.Get(ctx, hash)
+		st, err := db.Get(ctx, hash)
 		if err != nil {
 			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "read failed"})
 			return
@@ -69,12 +68,12 @@ func UserStateAPI(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "404 not found", http.StatusNotFound)
 			return
 		}
-		var body userstate.State
-		if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, maxUserStateBody)).Decode(&body); err != nil {
+		var body State
+		if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, maxBody)).Decode(&body); err != nil {
 			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "bad body"})
 			return
 		}
-		result, conflict, err := UserStateDB.Put(ctx, hash, body, body.Version)
+		result, conflict, err := db.Put(ctx, hash, body, body.Version)
 		if err != nil {
 			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "write failed"})
 			return
@@ -94,11 +93,11 @@ func UserStateAPI(w http.ResponseWriter, r *http.Request) {
 			Data    json.RawMessage `json:"data"`
 			Version int64           `json:"version"`
 		}
-		if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, maxUserStateBody)).Decode(&body); err != nil {
+		if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, maxBody)).Decode(&body); err != nil {
 			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "bad body"})
 			return
 		}
-		result, conflict, err := UserStateDB.Patch(ctx, hash, section, body.Data, body.Version)
+		result, conflict, err := db.Patch(ctx, hash, section, body.Data, body.Version)
 		if err != nil {
 			writeJSON(w, http.StatusBadRequest, map[string]string{"error": err.Error()})
 			return
